--- src/main/java/Command.java
@@ -1,4 +1,0 @@
-public abstract class Command {
-    public abstract void execute(TaskList taskList, Ui ui, Storage storage) throws LunaBotException;
-
-}
