+++ src/main/java/LunaBot/LunaBot.java
@@ -1,3 +1,13 @@
+package LunaBot;
+
+import LunaBot.command.Command;
+import LunaBot.command.ExitCommand;
+import LunaBot.storage.Storage;
+import LunaBot.ui.Ui;
+import LunaBot.task.TaskList;
+import LunaBot.parser.Parser;
+import LunaBot.exception.LunaBotException;
+
 import java.util.Scanner;
 
 public class LunaBot {
@@ -24,22 +34,21 @@ public class LunaBot {
         while (true) {  // Infinite loop, break on "bye" command
             try {
                 String input = ui.readCommand();
-                Command command = Parser.parse(input);  // Parse the input into a command
-                command.execute(taskList, ui, storage); // Execute the command on the taskList
+                Command command = Parser.parse(input);
+                command.execute(taskList, ui, storage);
 
-                // Check if the command is an ExitCommand, break the loop if so
+                // Check if the command is an LunaBot.command.ExitCommand, break the loop if so
                 if (command instanceof ExitCommand) {
-                    break;  // Exit the loop and end the program
+                    break;
                 }
             } catch (LunaBotException e) {
                 ui.printError(e.getMessage());
             }
         }
-        ui.printGoodbye();  // Show goodbye message after loop breaks
     }
 
     public static void main(String[] args) {
-        String filePath = "data/tasks.txt"; // Set the path for saving/loading tasks
-        new LunaBot(filePath).run();  // Run the LunaBot application
+        String filePath = "data/tasks.txt";
+        new LunaBot(filePath).run();
     }
 }
