+++ src/main/java/LunaBot/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package LunaBot.command;
+
+import LunaBot.storage.Storage;
+import LunaBot.task.TaskList;
+import LunaBot.ui.Ui;
+
 public class ExitCommand extends Command {
 
     @Override
