+++ src/main/java/LunaBot/command/MarkCommand.java
@@ -1,14 +1,22 @@
-public class DeleteCommand extends Command {
+package LunaBot.command;
+
+import LunaBot.exception.LunaBotException;
+import LunaBot.storage.Storage;
+import LunaBot.task.Task;
+import LunaBot.task.TaskList;
+import LunaBot.ui.Ui;
+
+public class MarkCommand extends Command {
     private int index;
 
-    public DeleteCommand(String input) throws LunaBotException {
+    public MarkCommand(String input) throws LunaBotException {
         try {
             // extracts index as a string and converts to an int
-            this.index = Integer.parseInt(input.substring(5)) - 1;
+            this.index = Integer.parseInt(input.substring(5).trim()) - 1;
         }
         catch (NumberFormatException e) {
             // checks if user inout an int
-            throw new LunaBotException(" Invalid task number format");
+            throw new LunaBotException("Invalid task number format");
         }
     }
 
@@ -18,8 +26,9 @@ public class DeleteCommand extends Command {
         if (index < 0 || index >= taskList.size()) {
             throw new LunaBotException("Invalid task number provided");
         }
-        Task deleted = taskList.deleteTask(index);
+        Task task = taskList.get(index);
+        task.markAsDone();
         storage.save(taskList.getTasks());
-        ui.printTaskDeleted(deleted, taskList.size());
+        ui.printTaskMarked(task);
     }
 }
