+++ src/main/java/LunaBot/exception/LunaBotException.java
@@ -1,3 +1,5 @@
+package LunaBot.exception;
+
 public class LunaBotException extends Exception {
     public LunaBotException(String message) {
         super(message);
