+++ src/main/java/LunaBot/task/Task.java
@@ -1,3 +1,5 @@
+package LunaBot.task;
+
 public abstract class Task {
     protected String description;
     protected boolean isDone;
