+++ src/main/java/LunaBot/task/ToDo.java
@@ -1,3 +1,5 @@
+package LunaBot.task;
+
 public class ToDo extends Task {
 
     public ToDo(String description, boolean isDone) {
