+++ src/main/java/LunaBot/ui/Ui.java
@@ -1,4 +1,7 @@
-import java.lang.reflect.Array;
+package LunaBot.ui;
+
+import LunaBot.task.Task;
+
 import java.util.ArrayList;
 import java.util.Scanner;
