+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ___________________________________________________________________
- Hello! I'm LunaBot
+ Hello! I'm LunaBot.LunaBot
  What can I do for you?
 ___________________________________________________________________
 ___________________________________________________________________
